+++ x/btcstaking/types/errors.go
@@ -27,10 +27,9 @@ var (
 	ErrInvalidProofOfPossession     = errorsmod.Register(ModuleName, 1119, "the proof of possession is not valid")
 	ErrDuplicatedFp                 = errorsmod.Register(ModuleName, 1120, "the staking request contains duplicated finality provider public key")
 	ErrInvalidBTCUndelegateReq      = errorsmod.Register(ModuleName, 1121, "invalid undelegation request")
-	ErrVotingPowerTableNotUpdated   = errorsmod.Register(ModuleName, 1122, "voting power table has not been updated")
-	ErrVotingPowerDistCacheNotFound = errorsmod.Register(ModuleName, 1123, "the voting power distribution cache is not found")
-	ErrParamsNotFound               = errorsmod.Register(ModuleName, 1124, "the parameters are not found")
-	ErrFpAlreadyJailed              = errorsmod.Register(ModuleName, 1125, "the finality provider has already been jailed")
-	ErrFpNotJailed                  = errorsmod.Register(ModuleName, 1126, "the finality provider is not jailed")
-	ErrDuplicatedCovenantSig        = errorsmod.Register(ModuleName, 1127, "the covenant signature is already submitted")
+	ErrVotingPowerDistCacheNotFound = errorsmod.Register(ModuleName, 1122, "the voting power distribution cache is not found")
+	ErrParamsNotFound               = errorsmod.Register(ModuleName, 1123, "the parameters are not found")
+	ErrFpAlreadyJailed              = errorsmod.Register(ModuleName, 1124, "the finality provider has already been jailed")
+	ErrFpNotJailed                  = errorsmod.Register(ModuleName, 1125, "the finality provider is not jailed")
+	ErrDuplicatedCovenantSig        = errorsmod.Register(ModuleName, 1126, "the covenant signature is already submitted")
 )
