+++ x/finality/abci.go
@@ -23,7 +23,7 @@ func EndBlocker(ctx context.Context, k keeper.Keeper) ([]abci.ValidatorUpdate, e
 
 	// if the BTC staking protocol is activated, i.e., there exists a height where a finality provider
 	// has voting power, start indexing and tallying blocks
-	if _, err := k.BTCStakingKeeper.GetBTCStakingActivatedHeight(ctx); err == nil {
+	if _, err := k.GetBTCStakingActivatedHeight(ctx); err == nil {
 		// index the current block
 		k.IndexBlock(ctx)
 		// tally all non-finalised blocks
